Re-apply changes to test_least_core_valuation.
